% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nitrate_clean}
\alias{nitrate_clean}
\title{Surface water nitrate concentrations from NEON sites}
\format{
\subsection{'nitrate_clean'}{

A data frame 602,889 rows and 7 columns
\describe{
\item{siteID}{NEON site identifier}
\item{surfWaterNitrateMean}{Mean nitrate concentration in surface water (µmol/L)}
\item{finalQF}{Final quality flag (0 = pass, 1 = fail)}
\item{startDate}{Start date of measurement (Date)}
\item{startTime}{Start time of measurement (character)}
\item{endDate}{End date of measurement (Date)}
\item{endTime}{End time of measurement (character)}
}
}
}
\source{
\url{https://data.neonscience.org/data-products/DP1.20033.001/RELEASE-2025}
}
\usage{
nitrate_clean
}
\description{
Surface water nitrate concentrations from NEON sites
}
\keyword{datasets}
