% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot distribution of nitrate concentrations}
\usage{
plot_distribution(
  data,
  sites = NULL,
  bin_size = 20,
  show_density = FALSE,
  facet_sites = FALSE
)
}
\arguments{
\item{data}{Filtered nitrate data}

\item{sites}{Selected site IDs}

\item{bin_size}{Number of bins for histogram}

\item{show_density}{Whether to show density curve}

\item{facet_sites}{Whether to facet by site}
}
\value{
plotly object showing distribution
}
\description{
Plot distribution of nitrate concentrations
}
\examples{
plot_distribution(nitrate_clean)
}
