% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nitrate_app.R
\name{run_nitrate_app}
\alias{run_nitrate_app}
\title{Run the Nitrate Explorer Shiny Application}
\usage{
run_nitrate_app()
}
\value{
None. This function runs the Shiny app.
}
\description{
Launches the Nitrate Explorer shiny application, which allows users to visualize and analyze Nitrate data.
}
\examples{
if (interactive()) {
  run_nitrate_app()
}
}
